+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.2.0+6 - 2026-06-08
+
+- Add script import from existing local script files.
+- Add custom vault storage folder selection with copy-before-switch behavior.
+
 ## 1.1.0+5 - 2026-06-07
 
 - TODO: Add release notes.
@@ -15,4 +20,3 @@
 ## 1.0.1+2 - 2026-06-02
 
 - TODO: Add release notes.
-